Adapt colour scheme for ZSH suggestions
ZSH suggestions present the last executed command that starts with the
current prefix. Here it is essential to be able to distinguish between
already entered letters and those only proposed. This was not the case
in the previous colour scheme.

This commit adapts some colours such that characters only suggested
appear in light grey and a floating auto-complete '/' appears in almost
black gray. A short usage has not revealed any conflicts, a longer use
still has to follow.