Make special highlighting visible
A CLI program printed some text using a highlight that happened to be
Color7. This was the same colour as the background, so the text was
invisible.

This commit changes Color7 to a darker grey. The exact tone was chosen
to be the "cursorline" color from the Vim colour scheme. By this, no new
colours are introduced to the palette. It also adapts Color0Intense to
the same colour in order to stay consistent.